backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.8.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-dotenv>=1.0.0
ccxt>=4.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
slowapi>=0.1.9
scipy>=1.10.0
scikit-learn>=1.3.0
pydantic-settings>=2.8.0
pytz>=2023.3
aiohttp>=3.8.0
python-telegram-bot>=20.0